% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgProgressBar.R
\docType{class}
\name{msgProgressBar-class}
\alias{msgProgressBar}
\alias{msgProgressBar-class}
\title{Message progress bar}
\description{
A simple progress bar to use in R packages where messages are prefered to console output.
}
\section{Fields}{

\describe{
\item{\code{iter}}{Total number of iterations}

\item{\code{i}}{Current iteration}

\item{\code{width}}{Width of the R console}

\item{\code{width_bar}}{Width of the progress bar}

\item{\code{progress}}{The number of character printed (continous)}

\item{\code{progress_step}}{Addition to progress per iteration}
}}
\section{Methods}{

\describe{
\item{\code{increment()}}{Increment the message bar iterator.}

\item{\code{initialize(iter)}}{Initialize a messagebar object}
}}
\examples{
test_bar <- function(i = 10){
 bar <- msgProgressBar(i)
 for(j in 1:i){
   bar$increment()
   Sys.sleep(4/i)
   }
 }
 test_bar(100)

}

